Move test in correct file
